Move LazyList to a separate module
